// SimERP/ClientApp/src/app/lists/product/model/productdetail.ts
import {ProductCIFPrice} from './productcifprice';

export class ProductDetail {
  Id = 0;
  ProductId = 0;
  ProductCode: string;
  SPKTName: string;
  ProductName: string;
  ProductNamePlan: string;
  ProductNameTKHQ: string;
  ProductNameFull: string;
  IdGroup: number;
  Description: string;
  IdUnit: number;
  Unit: string;
  Vatproduct: number;
  SupplierCode: string;
  Supplier: string;
  Producer: string;
  National: string;
  level: string;
  Application: string;
  pack: string;
  ConvertKg: number;
  GiaCifnorth: number;
  GiaCif: number;
  LoaiTien: string;
  ImportCode: string;
  ImportTax: number;
  CpveKhoLh: number;
  CpveKhoBn: number;
  Group: string;
  Salesman: string;
  LuongTonB: number;
  LuongTon: number;
  Leadtime: number;
  CtyPhuTrach: string;
  TgTonKhoTb: number;
  GhiChuTonKho: string;
  ViewNo: number;
  thutuxemB: number;
  thutuxem: number;
  FullFCL: number;
  KgPallet: number;
  ETAETD: number;
  CertNo: string;
  CertDate: Date;
  CertEnd: Date;
  Licenses: string;
  Remark: string;
  EmployeeId: number;
  SearchString: string;
  CreatedDate: Date;
  CreatedBy: number;
  ModifyDate: Date;
  ModifyBy: number;
  GiaMin: number = 0;
  IdKhmin: number;
  NgayGiaMin: Date;
  GiaMax: number = 0;
  IdKhmax: number;
  NgayGiaMax: Date;
  GiaTb: number = 0;
  LoaiTienNorth: string;

  // custom properties
  KhMinName: string;
  KhMaxName: string;
  ListProductCIFPrice: ProductCIFPrice[] = [];
  ListProductCIFPricsDel: ProductCIFPrice[] = [];
}