// SimERP/ClientApp/src/app/lists/customerdetail/model/customerdetail.ts
import { CustomerDeliveryAddress } from './customerdeliveryaddress';
import { Sales } from './Sales';

export class CustomerDetail {
  CusId: number;
  CustomerCode: string = '';
  NameCus: string = '';
  NameCusFS: string;
  TenTat: string;
  FullnameCus: string = '';
  NguoiMua: string;
  Khktoan: string;
  Adress: string;
  DeliveryAdr: string;
  MaKhuVuc: string;
  KhuVuc: string;
  DirectorName: string;
  ContactName: string;
  NgLienhe: string;
  LienHeCVu: string;
  NguoiKy: string;
  DtlienHe: string;
  Tel: string;
  Fax: string;
  Email: string;
  TaxCode: string;
  HanTt: any;
  SoTaiKhoan: string;
  Document: string;
  Giogiao: string;
  Type: string;
  Nvid: number;
  Manager: string;
  Employee: string;
  Na: boolean;
  DangBan: boolean;
  NoiGuiCongno: string;
  CCCongNo: string;
  CreatedDate: Date;
  CreatedBy: number;
  ModifyDate: Date;
  ModifyBy: number;
  IsActive: boolean = true;
  IsDelete: boolean = false;
  SearchString: string;
  SortOrder: number;
  PaymentExpired: number = 0;
  lstCustomerDeliveryAddress: CustomerDeliveryAddress[] = [];
  lstSales: Sales[] = [];
}